cleanup and making login, signup, dashboard, and form match styling

+++ apps/core/views.py
@@ -3,6 +3,8 @@ from django import forms
 from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib import auth
+from django.contrib.auth.decorators import login_required
+
 from py3wetransfer import Py3WeTransfer
 import os
 wetransfer_api_key = os.environ["WeTransfer_API_KEY"]
@@ -15,7 +17,7 @@ import datetime
 
 class QuietShareForm(forms.Form):
     text = forms.CharField(widget=forms.Textarea)
-    username = forms.CharField(max_length=30)
+    recipient_username = forms.CharField(max_length=30)
     filename = forms.FileField(label = "Select a file", required=False)
 
 # Two example views. Change or delete as necessary.
@@ -40,7 +42,7 @@ def home(request):
             expiry_date = dt_now + datetime.timedelta(days=7),
         )
 
-        return redirect('/')
+        return redirect('/accounts/login')
 
     else:
         # if a GET we'll create a blank form
@@ -52,57 +54,27 @@ def home(request):
 
     return render(request, 'pages/index.html', context)
 
+@login_required
 def user_page(request, username):
-    user = User.objects.get(username=username)
-    links = FilePost.objects.filter(username=user)
+    check = str(request.user)
+    if check == username:
+        user = User.objects.get(username=username)
+        links = FilePost.objects.filter(username=user)
 
-
-    if request.user == user:
-        is_viewing_self = True
+        context = {
+            'user': user,
+            'links': links,
+        }
+        return render(request, 'pages/user_page.html', context)
     else:
-        is_viewing_self = False
+        return redirect('/')
 
-    context = {
-        'user': user,
-        'links': links,
-    }
-    return render(request, 'pages/user_page.html', context)
 
+@login_required
 def delete_profile(request, username):
-    user = User.objects.get(username=username)
-    user.delete()
+    check = str(request.user)
+    if check == username:
+        user = User.objects.get(username=username)
+        user.delete()
 
     return redirect('/')
-
-# def file(request):
-#     if request.method == 'POST':
-#                 # Create a form instance and populate it with data from the request
-#         form = QuietShareForm(request.POST)
-#         x = Py3WeTransfer(wetransfer_api_key)
-#
-# #        if form.is_valid():
-#         filename=request.POST['filename']
-#         text=request.POST['text']
-#         link = x.upload_file(filename, text)
-#         dt_now = datetime.datetime.now()
-#
-#         filepost = FilePost.objects.create(
-#             username=request.POST['username'],
-#             text=text,
-#             link = link,
-#             expiry_date = dt_now + datetime.timedelta(days=7),
-#         )
-#
-#             # As soon as our new user is created, we make this user be
-#             # instantly "logged in"
-#             #auth.login(request, user)
-#         return redirect('/')
-    #
-    # else:
-    #     # if a GET we'll create a blank form
-    #     form = QuietShareForm()
-    #
-    # context = {
-    #     'form': form,
-    # }
-    # return render(request, 'pages/file0.html', context)
